requirements.txt
streamlit
pandas
numpy
bcrypt
plotly
fpdf2
openpyxl
email-validator
Pillow
python-dateutil
